deprecate exporter API in obsreport
The functionality has been moved into exporter helper. This is part of
