+++ obsreport/obsreport_test.go
@@ -28,7 +28,6 @@ var (
 	receiverID  = component.NewID("fakeReceiver")
 	scraperID   = component.NewID("fakeScraper")
 	processorID = component.NewID("fakeProcessor")
-	exporterID  = component.NewID("fakeExporter")
 
 	errFake        = errors.New("errFake")
 	partialErrFake = scrapererror.NewPartialScrapeError(errFake, 1)
@@ -255,152 +254,6 @@ func TestScrapeMetricsDataOp(t *testing.T) {
 		}
 
 		require.NoError(t, obsreporttest.CheckScraperMetrics(tt, receiverID, scraperID, int64(scrapedMetricPoints), int64(erroredMetricPoints)))
-	})
-}
-
-func TestExportTraceDataOp(t *testing.T) {
-	testTelemetry(t, exporterID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		parentCtx, parentSpan := tt.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
-		defer parentSpan.End()
-
-		obsrep, err := newExporter(ExporterSettings{
-			ExporterID:             exporterID,
-			ExporterCreateSettings: tt.ToExporterCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-
-		params := []testParams{
-			{items: 22, err: nil},
-			{items: 14, err: errFake},
-		}
-		for i := range params {
-			ctx := obsrep.StartTracesOp(parentCtx)
-			assert.NotNil(t, ctx)
-			obsrep.EndTracesOp(ctx, params[i].items, params[i].err)
-		}
-
-		spans := tt.SpanRecorder.Ended()
-		require.Equal(t, len(params), len(spans))
-
-		var sentSpans, failedToSendSpans int
-		for i, span := range spans {
-			assert.Equal(t, "exporter/"+exporterID.String()+"/traces", span.Name())
-			switch {
-			case params[i].err == nil:
-				sentSpans += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendSpansKey, Value: attribute.Int64Value(0)})
-				assert.Equal(t, codes.Unset, span.Status().Code)
-			case errors.Is(params[i].err, errFake):
-				failedToSendSpans += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentSpansKey, Value: attribute.Int64Value(0)})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendSpansKey, Value: attribute.Int64Value(int64(params[i].items))})
-				assert.Equal(t, codes.Error, span.Status().Code)
-				assert.Equal(t, params[i].err.Error(), span.Status().Description)
-			default:
-				t.Fatalf("unexpected error: %v", params[i].err)
-			}
-		}
-
-		require.NoError(t, tt.CheckExporterTraces(int64(sentSpans), int64(failedToSendSpans)))
-	})
-}
-
-func TestExportMetricsOp(t *testing.T) {
-	testTelemetry(t, exporterID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		parentCtx, parentSpan := tt.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
-		defer parentSpan.End()
-
-		obsrep, err := newExporter(ExporterSettings{
-			ExporterID:             exporterID,
-			ExporterCreateSettings: tt.ToExporterCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-
-		params := []testParams{
-			{items: 17, err: nil},
-			{items: 23, err: errFake},
-		}
-		for i := range params {
-			ctx := obsrep.StartMetricsOp(parentCtx)
-			assert.NotNil(t, ctx)
-
-			obsrep.EndMetricsOp(ctx, params[i].items, params[i].err)
-		}
-
-		spans := tt.SpanRecorder.Ended()
-		require.Equal(t, len(params), len(spans))
-
-		var sentMetricPoints, failedToSendMetricPoints int
-		for i, span := range spans {
-			assert.Equal(t, "exporter/"+exporterID.String()+"/metrics", span.Name())
-			switch {
-			case params[i].err == nil:
-				sentMetricPoints += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendMetricPointsKey, Value: attribute.Int64Value(0)})
-				assert.Equal(t, codes.Unset, span.Status().Code)
-			case errors.Is(params[i].err, errFake):
-				failedToSendMetricPoints += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentMetricPointsKey, Value: attribute.Int64Value(0)})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendMetricPointsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				assert.Equal(t, codes.Error, span.Status().Code)
-				assert.Equal(t, params[i].err.Error(), span.Status().Description)
-			default:
-				t.Fatalf("unexpected error: %v", params[i].err)
-			}
-		}
-
-		require.NoError(t, tt.CheckExporterMetrics(int64(sentMetricPoints), int64(failedToSendMetricPoints)))
-	})
-}
-
-func TestExportLogsOp(t *testing.T) {
-	testTelemetry(t, exporterID, func(t *testing.T, tt obsreporttest.TestTelemetry, useOtel bool) {
-		parentCtx, parentSpan := tt.TracerProvider.Tracer("test").Start(context.Background(), t.Name())
-		defer parentSpan.End()
-
-		obsrep, err := newExporter(ExporterSettings{
-			ExporterID:             exporterID,
-			ExporterCreateSettings: tt.ToExporterCreateSettings(),
-		}, useOtel)
-		require.NoError(t, err)
-
-		params := []testParams{
-			{items: 17, err: nil},
-			{items: 23, err: errFake},
-		}
-		for i := range params {
-			ctx := obsrep.StartLogsOp(parentCtx)
-			assert.NotNil(t, ctx)
-
-			obsrep.EndLogsOp(ctx, params[i].items, params[i].err)
-		}
-
-		spans := tt.SpanRecorder.Ended()
-		require.Equal(t, len(params), len(spans))
-
-		var sentLogRecords, failedToSendLogRecords int
-		for i, span := range spans {
-			assert.Equal(t, "exporter/"+exporterID.String()+"/logs", span.Name())
-			switch {
-			case params[i].err == nil:
-				sentLogRecords += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendLogRecordsKey, Value: attribute.Int64Value(0)})
-				assert.Equal(t, codes.Unset, span.Status().Code)
-			case errors.Is(params[i].err, errFake):
-				failedToSendLogRecords += params[i].items
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.SentLogRecordsKey, Value: attribute.Int64Value(0)})
-				require.Contains(t, span.Attributes(), attribute.KeyValue{Key: obsmetrics.FailedToSendLogRecordsKey, Value: attribute.Int64Value(int64(params[i].items))})
-				assert.Equal(t, codes.Error, span.Status().Code)
-				assert.Equal(t, params[i].err.Error(), span.Status().Description)
-			default:
-				t.Fatalf("unexpected error: %v", params[i].err)
-			}
-		}
-
-		require.NoError(t, tt.CheckExporterLogs(int64(sentLogRecords), int64(failedToSendLogRecords)))
 	})
 }
